+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added intel.handlers.Rotating
 - added ignore options to intel.console()
+- added chalk.enabled when colorize is used
 
 ## v0.2.0 - 2013-10-04